fixed panic error (first place I looked!) experimenting with simplifying stuff in `/net` so you can use `go run` if one wants
