--- net/client.go
@@ -1,57 +1,0 @@
-package main
-
-import (
-	"time"
-
-	"github.com/benjamin-rood/abm-colour-polymorphism/abm"
-
-	"golang.org/x/net/websocket"
-)
-
-// Client defines the owner of the Model.
-type Client struct {
-	*websocket.Conn
-	UID string
-	abm.Model
-	Active bool
-	Stamp  time.Time
-	Quit   chan struct{}
-}
-
-// NewClient constructs an initialised Client session.
-func NewClient(ws *websocket.Conn, uid string) (c Client) {
-	c.Conn = ws
-	c.UID = uid
-	c.Model = abm.NewModel()
-	c.Active = true
-	c.Stamp = time.Now()
-	c.Quit = make(chan struct{})
-	return
-}
-
-// Monitor keeps the client's connection alive,
-// and responds to any internal running model signaling
-// – e.g. if there is a fault in the running abm,
-// or if the population of the CP Prey agents reaches zero,
-// then the model will invoke Kill() and Quit will close,
-// which permits us to clean up and disconnect the Client.
-func (c *Client) Monitor(ch chan struct{}) {
-	defer func() {
-		c.Active = false
-		c.Stamp = time.Now()
-	}()
-	for {
-		select {
-		case <-c.Quit: //	internal signal from client.
-			close(ch) //	exit websocket connection.
-			// send final statistics
-			// clean up
-			return
-		case <-ch:
-			c.Suspend() //	websocket connection dead, suspend model operation.
-			return
-		default:
-			time.Sleep(time.Millisecond * 100)
-		}
-	}
-}
